refactor: custom hook file
