--- src/modules/dashboard/Board/shared/hook/useDialogBoard/types/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./useDialogBoard.types";
-export * from "./useDialogBoard.types";
