+++ src/modules/dashboard/Board/shared/hooks/useDialogBoard/index.ts
