+++ src/modules/dashboard/Board/shared/hooks/useDialogBoard/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-dialog-board.types";
