+++ src/modules/dashboard/Board/shared/hooks/useDialogBoard/types/use-dialog-board.types.ts
